Good ol Country Refactor
