+++ matico_components/src/Components/MaticoEditor/EditorComponents/DatasetTransformEditor/ColumnTransfomStepEditor.tsx
@@ -1,29 +1,17 @@
 import React from "react";
 import {
-    DialogTrigger,
     ActionButton,
-    Dialog,
-    Content,
     Flex,
     Picker,
     Item,
-    NumberField,
-    ToggleButton,
     Divider,
     Text
 } from "@adobe/react-spectrum";
-import { VariableSelector } from "Components/MaticoEditor/Utils/VariableSelector";
 import {
-    FilterStep,
-    Filter,
     ColumnTransformStep,
-    DateOpts,
-    FloatOpts,
-    IntOpts,
     ColumnTransform
+    // @ts-ignore
 } from "@maticoapp/matico_types/spec";
-import FunctionIcon from "@spectrum-icons/workflow/Function";
-import { FilterEditor } from "Components/MaticoEditor/Utils/FilterEditor";
 import { DatasetColumnSelector } from "Components/MaticoEditor/Utils/DatasetColumnSelector";
 import {Column} from "Datasets/Dataset";
 
@@ -34,7 +22,7 @@ interface ColumnTransformEditorProps {
     onUpdateTransform: (update: Partial<ColumnTransform>) => void;
 }
 
-const ColumnTransformEditor: React.FC<ColumnTransformEditorProps> = ({datasetId,transfom,columns, onUpdateTransform}) => {
+const ColumnTransformEditor: React.FC<ColumnTransformEditorProps> = ({datasetId,transform,columns, onUpdateTransform}) => {
     console.log("transform ", transform, " Columns ",columns)
     return (
         <Flex>
