+++ matico_components/src/Components/MaticoEditor/EditorComponents/DatasetTransformEditor/FilterStep/FilterTypeDialog.tsx
@@ -1,0 +1,64 @@
+import React from "react";
+import {
+    DialogTrigger,
+    ActionButton,
+    Dialog,
+    Content,
+    Flex
+} from "@adobe/react-spectrum";
+
+export const FilterTypeDialog: React.FC<{ onSubmit: (newFilter: any) => void }> = ({
+    onSubmit
+}) => {
+    return (
+        <DialogTrigger isDismissable type="popover">
+            <ActionButton>Add Filter</ActionButton>
+            {(close) => (
+                <Dialog>
+                    <Content>
+                        <Flex direction="column" gap={"size-200"}>
+                            <ActionButton
+                                onPress={() => {
+                                    onSubmit({
+                                        variable: null,
+                                        min: 0,
+                                        max: 100,
+                                        type: "range"
+                                    });
+                                    close();
+                                }}
+                            >
+                                Range
+                            </ActionButton>
+                            <ActionButton
+                                onPress={() => {
+                                    onSubmit({
+                                        variable: null,
+                                        min: new Date(2000, 1, 1),
+                                        max: new Date(2022, 1, 1),
+                                        type: "date"
+                                    });
+                                    close();
+                                }}
+                            >
+                                Date Range
+                            </ActionButton>
+                            <ActionButton
+                                onPress={() => {
+                                    onSubmit({
+                                        variable: null,
+                                        is_one_of: [],
+                                        type: "category"
+                                    });
+                                    close();
+                                }}
+                            >
+                                Category
+                            </ActionButton>
+                        </Flex>
+                    </Content>
+                </Dialog>
+            )}
+        </DialogTrigger>
+    );
+};
