+++ matico_components/src/Components/MaticoEditor/EditorComponents/DatasetTransformEditor/FilterStep/index.ts
@@ -1,0 +1,13 @@
+import { CategoryFilter } from "./CategoryFilter";
+import { DateRangeFilterEditor } from "./DateRangeFilterEditor";
+import { FilterStepEditor } from "./FilterStep";
+import { FilterTypeDialog } from "./FilterTypeDialog";
+import { RangeFilterEditor } from "./RangeFilterEditor";
+
+export {
+    CategoryFilter,
+    DateRangeFilterEditor,
+    FilterTypeDialog,
+    RangeFilterEditor,
+    FilterStepEditor
+};
