+++ matico_components/src/Components/MaticoEditor/EditorComponents/DatasetTransformEditor/index.ts
@@ -1,0 +1,4 @@
+import { DatasetTransformEditor } from "./DatasetTransformEditor";
+import { DatasetTransformDialog } from "./DatasetTransformDialog";
+
+export { DatasetTransformEditor, DatasetTransformDialog };
