+++ matico_components/src/Components/MaticoEditor/EditorComponents/DatasetTransformPane/DatasetTransformPane.tsx
@@ -4,20 +4,16 @@ import {
     Flex,
     Heading,
     TextField,
-    Well,
-    Text,
     Content,
     Dialog,
     TextArea
 } from "@adobe/react-spectrum";
-import { Dataset, DatasetTransform } from "@maticoapp/matico_types/spec";
+// @ts-ignore
+import { DatasetTransform } from "@maticoapp/matico_types/spec";
 import { DatasetSelector } from "Components/MaticoEditor/Utils/DatasetSelector";
 import { useApp } from "Hooks/useApp";
 import React, { useState } from "react";
-import {
-    DatasetTransformDialog,
-    DatasetTransformEditor
-} from "../DatasetTransformEditor/DatasetTransformEditor";
+import { DatasetTransformDialog } from "../DatasetTransformEditor";
 import { v4 as uuid } from "uuid";
 import { useMaticoSelector } from "Hooks/redux";
 
@@ -99,7 +95,7 @@ export const DatasetTransfromPane: React.FC = () => {
             </Heading>
 
             <Flex direction="column">
-                {datasetTransforms?.map((transform) => (
+                {datasetTransforms?.map((transform: any) => (
                     <DatasetTransformDialog
                         key={transform.id}
                         transformId={transform.id}
